feat: 공통적으로 사용되는 리소스에 대한 타입을 types.ts 파일로 분리
