+++ src/api/types.ts
@@ -1,0 +1,32 @@
+export interface Category {
+	code: number;
+	main: string;
+	sub: string;
+}
+
+export interface Host {
+	email: string;
+	id: string;
+	isLogout: boolean;
+	password: string;
+	role: string;
+	token: string;
+}
+
+export interface Study {
+	HOST_ID: string;
+	capacity: number;
+	categoryCode?: Category;
+	createdAt: string;
+	frequency: string;
+	hostId?: Host;
+	id: string;
+	isOpen: boolean;
+	location: string;
+	membersCount: number;
+	studyAbout: string;
+	title: string;
+	vacancy: number;
+	views: number;
+	weekday: string;
+}
